+++ Model/Method/Adapter.php
@@ -21,6 +21,7 @@ use Magento\Payment\Gateway\Config\ValueHandlerPoolInterface;
 use Magento\Payment\Gateway\Validator\ValidatorPoolInterface;
 use Magento\Quote\Api\Data\PaymentInterface;
 
+// TODO: This class can probably be deleted since its use will be done in Model/AdyenPaymentMethod.php
 class Adapter extends Method\Adapter
 {
     /**
