+++ etc/di.xml
@@ -83,24 +83,14 @@
             <argument name="commandPool" xsi:type="object">AdyenPaymentHppCommandPool</argument>
         </arguments>
     </virtualType>
-    <virtualType name="AdyenPaymentIdealFacade" type="Adyen\Payment\Model\Method\Adapter">
+    <type name="Adyen\Payment\Model\AdyenPaymentMethod">
         <arguments>
-            <argument name="code" xsi:type="const">Adyen\Payment\Model\Ui\AdyenPaymentMethodConfigProvider::CODE_IDEAL</argument>
             <argument name="infoBlockType" xsi:type="string">Magento\Payment\Block\ConfigurableInfo</argument>
             <argument name="valueHandlerPool" xsi:type="object">AdyenPaymentHppValueHandlerPool</argument>
             <argument name="validatorPool" xsi:type="object">AdyenPaymentHppValidatorPool</argument>
             <argument name="commandPool" xsi:type="object">AdyenPaymentHppCommandPool</argument>
         </arguments>
-    </virtualType>
-    <virtualType name="AdyenPaymentKlarnaFacade" type="Adyen\Payment\Model\Method\Adapter">
-        <arguments>
-            <argument name="code" xsi:type="const">Adyen\Payment\Model\Ui\AdyenPaymentMethodConfigProvider::CODE_KLARNA</argument>
-            <argument name="infoBlockType" xsi:type="string">Magento\Payment\Block\ConfigurableInfo</argument>
-            <argument name="valueHandlerPool" xsi:type="object">AdyenPaymentHppValueHandlerPool</argument>
-            <argument name="validatorPool" xsi:type="object">AdyenPaymentHppValidatorPool</argument>
-            <argument name="commandPool" xsi:type="object">AdyenPaymentHppCommandPool</argument>
-        </arguments>
-    </virtualType>
+    </type>
     <virtualType name="AdyenPaymentBoletoFacade" type="Magento\Payment\Model\Method\Adapter">
         <arguments>
             <argument name="code" xsi:type="const">Adyen\Payment\Model\Ui\AdyenBoletoConfigProvider::CODE</argument>
